Fixed the issues

+++ PFF2 Team Project/Assets/Scripts/WandStats.cs
@@ -7,7 +7,7 @@ public class WandStats : ScriptableObject
 {
     public GameObject model;
    
-    [Range(1, 10)] public int shootDamageMod;   
+    [Range(1, 10)] public float shootDamageMod;   
     [Range(0.1f, 3)] public float shootRate;
 
     public ParticleSystem hitEffect;
